backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
jinja2==3.1.2
itsdangerous==2.1.2
starlette==0.27.0
typing-extensions==4.8.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pandas==2.1.0
numpy==1.25.2
prophet==1.1.4
scikit-learn==1.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.0
PyPDF2==3.0.1
google-generativeai==0.3.1 